refactor: split nor flow tests
