+++ lib/index.ts
@@ -11,6 +11,7 @@ export * from "./event";
 export * from "./keccak";
 export * from "./log";
 export * from "./network";
+export * from "./nor";
 export * from "./oracle";
 export * from "./promise";
 export * from "./proxy";
