+++ test/0.4.24/nor/nor.signing.keys.test.ts
@@ -1,0 +1,153 @@
+import { expect } from "chai";
+import { encodeBytes32String, ZeroAddress } from "ethers";
+import { ethers } from "hardhat";
+
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
+import { time } from "@nomicfoundation/hardhat-network-helpers";
+
+import {
+  ACL,
+  Kernel,
+  Lido,
+  LidoLocator,
+  LidoLocator__factory,
+  NodeOperatorsRegistry__MockForFlow,
+  NodeOperatorsRegistry__MockForFlow__factory,
+} from "typechain-types";
+
+import { certainAddress, ether, randomAddress } from "lib";
+
+import { addAragonApp, deployLidoDao } from "test/deploy";
+import { Snapshot } from "test/suite";
+
+describe("NodeOperatorsRegistry", () => {
+  let deployer: HardhatEthersSigner;
+  let user: HardhatEthersSigner;
+  let stranger: HardhatEthersSigner;
+
+  let limitsManager: HardhatEthersSigner;
+  let nodeOperatorsManager: HardhatEthersSigner;
+  let signingKeysManager: HardhatEthersSigner;
+  let stakingRouter: HardhatEthersSigner;
+  let lido: Lido;
+  let dao: Kernel;
+  let acl: ACL;
+  let locator: LidoLocator;
+
+  let impl: NodeOperatorsRegistry__MockForFlow;
+  let nor: NodeOperatorsRegistry__MockForFlow;
+
+  let originalState: string;
+
+  const NODE_OPERATORS: NodeOperatorConfig[] = [
+    {
+      name: "foo",
+      rewardAddress: certainAddress("node-operator-1"),
+      totalSigningKeysCount: 10n,
+      depositedSigningKeysCount: 5n,
+      exitedSigningKeysCount: 1n,
+      vettedSigningKeysCount: 6n,
+      stuckValidatorsCount: 0n,
+      refundedValidatorsCount: 0n,
+      stuckPenaltyEndAt: 0n,
+    },
+    {
+      name: " bar",
+      rewardAddress: certainAddress("node-operator-2"),
+      totalSigningKeysCount: 15n,
+      depositedSigningKeysCount: 7n,
+      exitedSigningKeysCount: 0n,
+      vettedSigningKeysCount: 10n,
+      stuckValidatorsCount: 0n,
+      refundedValidatorsCount: 0n,
+      stuckPenaltyEndAt: 0n,
+    },
+    {
+      name: "deactivated",
+      isActive: false,
+      rewardAddress: certainAddress("node-operator-3"),
+      totalSigningKeysCount: 10n,
+      depositedSigningKeysCount: 0n,
+      exitedSigningKeysCount: 0n,
+      vettedSigningKeysCount: 5n,
+      stuckValidatorsCount: 0n,
+      refundedValidatorsCount: 0n,
+      stuckPenaltyEndAt: 0n,
+    },
+  ];
+
+  const moduleType = encodeBytes32String("curated-onchain-v1");
+  const penaltyDelay = 86400n;
+  const contractVersion = 2n;
+
+  before(async () => {
+    [deployer, user, stakingRouter, nodeOperatorsManager, signingKeysManager, limitsManager, stranger] =
+      await ethers.getSigners();
+
+    ({ lido, dao, acl } = await deployLidoDao({
+      rootAccount: deployer,
+      initialized: true,
+      locatorConfig: {
+        stakingRouter,
+      },
+    }));
+
+    impl = await new NodeOperatorsRegistry__MockForFlow__factory(deployer).deploy();
+    const appProxy = await addAragonApp({
+      dao,
+      name: "node-operators-registry",
+      impl,
+      rootAccount: deployer,
+    });
+
+    nor = NodeOperatorsRegistry__MockForFlow__factory.connect(appProxy, deployer);
+
+    await acl.createPermission(user, lido, await lido.RESUME_ROLE(), deployer);
+
+    await acl.createPermission(stakingRouter, nor, await nor.STAKING_ROUTER_ROLE(), deployer);
+    await acl.createPermission(signingKeysManager, nor, await nor.MANAGE_SIGNING_KEYS(), deployer);
+    await acl.createPermission(nodeOperatorsManager, nor, await nor.MANAGE_NODE_OPERATOR_ROLE(), deployer);
+    await acl.createPermission(limitsManager, nor, await nor.SET_NODE_OPERATOR_LIMIT_ROLE(), deployer);
+
+    // grant role to nor itself cause it uses solidity's call method to itself
+    // inside the testing_requestValidatorsKeysForDeposits() method
+    await acl.grantPermission(nor, nor, await nor.STAKING_ROUTER_ROLE());
+
+    locator = LidoLocator__factory.connect(await lido.getLidoLocator(), user);
+
+    // Initialize the nor's proxy.
+    await expect(nor.initialize(locator, moduleType, penaltyDelay))
+      .to.emit(nor, "ContractVersionSet")
+      .withArgs(contractVersion)
+      .and.to.emit(nor, "LocatorContractSet")
+      .withArgs(locator)
+      .and.to.emit(nor, "StakingModuleTypeSet")
+      .withArgs(moduleType);
+
+    nor = nor.connect(user);
+  });
+
+  beforeEach(async () => (originalState = await Snapshot.take()));
+
+  afterEach(async () => await Snapshot.restore(originalState));
+
+  context("addSigningKeys", () => {});
+
+  context("addSigningKeysOperatorBH", () => {});
+
+  context("removeSigningKey", () => {});
+
+  context("removeSigningKeys", () => {});
+
+  context("removeSigningKeyOperatorBH", () => {});
+
+  context("removeSigningKeysOperatorBH", () => {});
+
+  context("getTotalSigningKeyCount", () => {});
+
+  context("getUnusedSigningKeyCount", () => {});
+
+  context("getSigningKey", () => {});
+
+  context("getSigningKeys", () => {});
+});
